requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
openai>=1.30.0
python-dotenv>=1.0.0
urllib3>=2.0.0
